Refactor all the code
