+++ student.rb
@@ -1,11 +1,12 @@
 require_relative 'person'
 
 class Student < Person
-  attr_reader :classroom
+  attr_reader :classroom, :parent_permission
 
   def initialize(age, classroom, name = 'Unknown', parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
+    super(age, name)
     @classroom = classroom
+    @parent_permission = parent_permission
     classroom&.add_student(self)
   end
